requirements.txt
asgiref==3.7.2
Django==4.2.6
django-appconf==1.0.5
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg==3.1.8
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
